Level Loading
I've implemented level loading. Now you can load levels from a file.
First the existing level gets cleared, then the program loops through
every line of the level file and converts it to a string and then trims
the newline character at the end. Then it creates the right tile
(determinded by the character in the file) at the right position
(determinded by the position in the file itself).